limit fast-sync  rate

+++ sunflower-core/src/main/java/org/tdf/sunflower/SyncConfig.java
@@ -58,5 +58,5 @@ public class SyncConfig {
     }
 
     @Getter
-    private Map<String, Integer> rateLimits;
+    private Map<String, Double> rateLimits;
 }
